+++ src/FileScannerPCRE2.cpp
@@ -21,15 +21,19 @@
 
 #include "FileScannerPCRE2.h"
 
-#include <iostream>
 #include <future/string.hpp>
+#include <libext/string.hpp>
+#include <libext/exception.hpp>
+
+#include <iostream>
 #include <libext/Logger.h>
 #include <cstring>
 
-#include "libext/hints.hpp"
+#include <libext/hints.hpp>
+#include <libext/memory.hpp>
 
 
-#ifdef HAVE_LIBPCRE2
+#if HAVE_LIBPCRE2
 /**
  * This callout handler is invoked by PCRE2 at the end of a potentially successful match.  It's purpose
  * is to prevent a regex like 'abc\s+def' from matching across an eol boundary, since '\s' matches both
@@ -64,7 +68,7 @@ static int callout_handler(pcre2_callout_block *cob, void *ctx)
 	}
 }
 
-static int count_callouts_callback([[maybe_unused]] pcre2_callout_enumerate_block *ceb, void *ctx)
+static int count_callouts_callback(pcre2_callout_enumerate_block *ceb [[maybe_unused]], void *ctx)
 {
 	size_t * ctr { reinterpret_cast<size_t*>(ctx) };
 
@@ -96,11 +100,12 @@ FileScannerPCRE2::FileScannerPCRE2(sync_queue<FileID> &in_queue,
 		bool word_regexp,
 		bool pattern_is_literal) : FileScanner(in_queue, output_queue, regex, ignore_case, word_regexp, pattern_is_literal)
 {
-#ifdef HAVE_LIBPCRE2
+#if HAVE_LIBPCRE2
 	// Compile the regex.
 	int error_code;
 	PCRE2_SIZE error_offset;
 	uint32_t regex_compile_options = 0;
+	std::string original_pattern = regex;
 
 	// For now, we won't support capturing.  () will be treated as (?:).
 	regex_compile_options = PCRE2_NO_AUTO_CAPTURE | PCRE2_MULTILINE | PCRE2_NEVER_BACKSLASH_C |  PCRE2_NEVER_UTF | PCRE2_NEVER_UCP
@@ -109,6 +114,7 @@ FileScannerPCRE2::FileScannerPCRE2(sync_queue<FileID> &in_queue,
 	if(ignore_case)
 	{
 		// Ignore case while matching.
+		LOG(INFO) << "Ignoring case.";
 		regex_compile_options |= PCRE2_CASELESS;
 	}
 
@@ -161,17 +167,93 @@ FileScannerPCRE2::FileScannerPCRE2(sync_queue<FileID> &in_queue,
 		pcre2_code_free(m_pcre2_regex);
 		throw FileScannerException(std::string("Callouts not supported."));
 	}
+
+	// Do our own analysis and see if there's anything we can do to help speed up the matching.
+	AnalyzeRegex(original_pattern);
+
 #endif
 }
 
 FileScannerPCRE2::~FileScannerPCRE2()
 {
-#ifdef HAVE_LIBPCRE2
+#if HAVE_LIBPCRE2
 	pcre2_code_free(m_pcre2_regex);
 #endif
 }
 
-#ifdef HAVE_LIBPCRE2
+std::string FileScannerPCRE2::GetPCRE2Version() noexcept
+{
+#if HAVE_LIBPCRE2
+
+	/// @note A bug in the PCRE2 docs says the PCRE2_CONFIG_VERSION buffer needs to be 13 code units long.
+	/// That isn't correct.  The version will come back as e.g. "10.22 2016-07-29", which is 17 chars including
+	/// the terminating '\0'.
+	return to_string(pcre2_config, PCRE2_CONFIG_VERSION);
+#else
+	return "none";
+#endif
+}
+
+void FileScannerPCRE2::AnalyzeRegex(const std::string &regex_passed_in) noexcept
+{
+#if HAVE_LIBPCRE2 == 0
+	(void)regex_passed_in;
+#else
+	// Check for a static first code unit or units.
+
+	// Check for a first code unit bitmap.
+	const uint8_t *first_bitmap {nullptr};
+	pcre2_pattern_info(m_pcre2_regex, PCRE2_INFO_FIRSTBITMAP, &first_bitmap);
+	if(first_bitmap != nullptr)
+	{
+		ConstructCodeUnitTable(first_bitmap);
+		m_use_find_first_of = true;
+		LOG(INFO) << "First code unit of pattern is one of '" << std::string((const char*)m_compiled_cu_bitmap, m_end_index) << "'.";
+	}
+	else
+	{
+		uint32_t first_code_type {0};
+		pcre2_pattern_info(m_pcre2_regex, PCRE2_INFO_FIRSTCODETYPE, &first_code_type);
+		if(first_code_type == 1 && !m_ignore_case) /// @todo This looks like a bug in PCRE2. If PCRE2_CASELESS is given, you will still get
+													/// a single FIRSTCODEUNIT back which has the case you passed in.  Same bug does not
+													/// seem to exist for the BITMAP; e.g. [A-Z] comes back as A-Za-z.
+		{
+			// There is a singular first code unit.
+			uint32_t first_code_unit {0};
+			pcre2_pattern_info(m_pcre2_regex, PCRE2_INFO_FIRSTCODEUNIT, &first_code_unit);
+			m_compiled_cu_bitmap[0] = first_code_unit;
+			m_end_index = 1;
+			m_use_find_first_of = true;
+			LOG(INFO) << "First code unit of pattern is '" << m_compiled_cu_bitmap << "'.";
+
+		}
+		else if(first_code_type == 2)
+		{
+			// The "first code unit" is the start of any line.
+			/// @todo Not sure we can make good use of this.
+		}
+	}
+
+	// If we have a static first code unit, let's check and see if the string is not a regex but a literal.
+	auto pat_is_lit = IsPatternLiteral(regex_passed_in);
+	if(!m_ignore_case              // If we're not ignoring case (smart case set this in ArgParse, @todo probably should move)...
+			&& !m_word_regexp                   // ... and we aren't doing a --word-regexp...
+			&& (m_pattern_is_literal || pat_is_lit))  // And we've been told to treat the pattern as literal, or it actually is literal
+	{
+		// This is a simple string comparison, we can bypass libpcre2 entirely.
+
+		constexpr auto vec_size_bytes = 16;
+
+		LOG(INFO) << "Using caseful literal search optimization";
+		m_literal_search_string_len = regex_passed_in.size();
+		size_t size_to_alloc = m_literal_search_string_len+1;
+		m_literal_search_string.reset(static_cast<uint8_t*>(overaligned_alloc(vec_size_bytes, size_to_alloc)));
+		std::memcpy(static_cast<void*>(m_literal_search_string.get()), static_cast<const void*>(regex_passed_in.c_str()), size_to_alloc);
+	}
+#endif
+}
+
+#if HAVE_LIBPCRE2
 /// @name Custom deleters for the PCRE2 objects we'll be using.
 /// These are implemented as specializations of the std::default_delete<> template.
 /// @{
@@ -196,7 +278,9 @@ struct default_delete<pcre2_match_context>
 
 void FileScannerPCRE2::ScanFile(const char* __restrict__ file_data, size_t file_size, MatchList& ml)
 {
-#ifdef HAVE_LIBPCRE2
+#if HAVE_LIBPCRE2
+	try
+	{
 	// Pointer to the offset vector returned by pcre2_match().
 	PCRE2_SIZE *ovector;
 
@@ -238,16 +322,42 @@ void FileScannerPCRE2::ScanFile(const char* __restrict__ file_data, size_t file_
 			options = PCRE2_NOTEMPTY_ATSTART | PCRE2_ANCHORED;
 		}
 
-		// Try to match the regex to whatever's left of the file.
-		int rc = pcre2_match(
-				m_pcre2_regex,
-				reinterpret_cast<PCRE2_SPTR>(file_data),
-				file_size,
-				start_offset,
-				options,
-				match_data.get(),
-				mctx.get()
-				);
+		/////
+		if(false)//m_use_find_first_of)
+		{
+			// Burn through chars we know aren't at the start of the match.
+			auto first_possible_char = FindFirstPossibleCodeUnit_default(file_data+start_offset, file_size-start_offset);
+			if(first_possible_char != file_data+file_size)
+			{
+				// Found one.
+				start_offset = first_possible_char - file_data;
+			}
+			else
+			{
+				// Found nothing, the regex can't match.
+				break;
+			}
+		}
+		///////
+
+		int rc = 0;
+		if(m_literal_search_string_len == 0)
+		{
+			// Try to match the regex to whatever's left of the file.
+			rc = pcre2_match(
+					m_pcre2_regex,
+					reinterpret_cast<PCRE2_SPTR>(file_data),
+					file_size,
+					start_offset,
+					options,
+					match_data.get(),
+					mctx.get()
+					);
+		}
+		else
+		{
+			rc = LiteralMatch(this, file_data, file_size, start_offset, ovector);
+		}
 
 		// Check for no match.
 		if(rc == PCRE2_ERROR_NOMATCH)
@@ -319,18 +429,37 @@ void FileScannerPCRE2::ScanFile(const char* __restrict__ file_data, size_t file_
 			return;
 		}
 
-		// There was a match.  Package it up in the MatchList which was passed in.
-		line_no += CountLinesSinceLastMatch(prev_lineno_search_end, file_data+ovector[0]);
-		prev_lineno_search_end = file_data+ovector[0];
-		if(line_no == prev_lineno)
+		try
 		{
-			// Skip multiple matches on one line.
-			continue;
-		}
-		prev_lineno = line_no;
-		Match m(file_data, file_size, ovector[0], ovector[1], line_no);
+			// There was a match.  Package it up in the MatchList which was passed in.
+			line_no += CountLinesSinceLastMatch(prev_lineno_search_end, file_data+ovector[0]);
+			prev_lineno_search_end = file_data+ovector[0];
+			if(line_no == prev_lineno)
+			{
+				// Skip multiple matches on one line.
+				continue;
+			}
+			prev_lineno = line_no;
+			Match m(file_data, file_size, ovector[0], ovector[1], line_no);
 
-		ml.AddMatch(std::move(m));
+			ml.AddMatch(std::move(m));
+		}
+		catch(...)
+		{
+			RETHROW("file_size=" + std::to_string(file_size) + ", ovector[0]=" + std::to_string(ovector[0])
+				+ ", ovector[1]=" + std::to_string(ovector[1])
+				+ ", start_offset=" + std::to_string(start_offset)
+			);
+		}
+	}
+	}
+	catch(const std::exception& e)
+	{
+		print_exception_stack(e);
+	}
+	catch(...)
+	{
+		RETHROW("Caught exception here.");
 	}
 #endif // HAVE_LIBPCRE2
 }
@@ -338,7 +467,7 @@ void FileScannerPCRE2::ScanFile(const char* __restrict__ file_data, size_t file_
 std::string FileScannerPCRE2::PCRE2ErrorCodeToErrorString(int errorcode)
 {
 	std::string retstr = "";
-#ifdef HAVE_LIBPCRE2
+#if HAVE_LIBPCRE2
 	PCRE2_SIZE error_msg_buf_size = 512;
 	PCRE2_UCHAR *error_msg_buf = new PCRE2_UCHAR[error_msg_buf_size];
 	int retval = pcre2_get_error_message(errorcode, error_msg_buf, error_msg_buf_size);