+++ src/future/dummy.cpp
@@ -24,7 +24,7 @@
 #include "memory.hpp"
 
 
-#include <static_diagnostics.hpp>
+#include "memory.hpp"
 
 const char *link_me = "dummy";