+++ src/libext/string.hpp
@@ -80,15 +80,10 @@ typename ContainerType::value_type join(const ContainerType& container_of_string
 	// size of the resulting string we'll end up with so we can reserve it, eliminating any reallocations.
 	// The second scan then does the actual join of the contained strings into this preallocated space.
 
-	// In C++17, this could be std::reduce() instead for a possible performance improvement (no order guarantee).
-	// Would likely be minimal though unless you were joining millions of strings.
-	size_t len = std::accumulate(container_of_strings.cbegin(), container_of_strings.cend(), 0,
-			[&separator](size_t accumulator, const typename ContainerType::value_type &str)
-				{
-					size_t retval = accumulator == 0 ? 0 : separator.length();
-					retval += str.length();
-					return retval;
-				});
+	size_t len = 0;
+	size_t seplen = separator.length();
+	std::for_each(container_of_strings.cbegin(), container_of_strings.cend(),
+			[&len, seplen](const typename ContainerType::value_type &str){ len += seplen + str.length();});
 	len += 1; // For any trailing '\0'.
 
 	// Create the return value and allocate the space we'll need.
@@ -123,6 +118,30 @@ std::string to_string(T val, std::ios_base & (*base)(std::ios_base&))
 	ss << std::showbase << std::setw(sizeof(T)*2+2) << base << val;
 
 	return ss.str();
+}
+
+
+/**
+ * Wrapper for those (mostly C) functions which take a buffer which they will fill with a string of unknown
+ * length, making you guess at how large a buffer you need.  This template papers over that nonsense.
+ * Requires #func to have the following properties:
+ * - Must take two parameters, the first being a "selector" of some type for selecting which string you want, the second
+ *   being a pointer to a buffer.
+ * - Must return the necessary length of the buffer when passed #tellmethelength as its second parameter.
+ *
+ * @param func
+ * @param s
+ * @return  A std:string containing the string #func put in the temporary buffer.
+ */
+template <typename CallableTwoParam, typename SelectorType, typename BuffType = char, BuffType *tellmethelength = nullptr>
+std::string to_string(CallableTwoParam func, SelectorType s)
+{
+	auto len = func(s, tellmethelength);
+	auto buffer = new BuffType[len+1];
+	len = func(s, buffer);
+	std::string retval {buffer};
+	delete [] buffer;
+	return retval;
 }
 
 #endif /* SRC_LIBEXT_STRING_HPP_ */